+++ tower-http/src/trace/make_span.rs
@@ -1,5 +1,7 @@
 use http::Request;
-use tracing::Span;
+use tracing::{Level, Span};
+
+use super::DEFAULT_MESSAGE_LEVEL;
 
 /// Trait used to generate [`Span`]s from requests. [`Trace`] wraps all request handling in this
 /// span.
@@ -30,8 +32,9 @@ where
 ///
 /// [`Span`]: tracing::Span
 /// [`Trace`]: super::Trace
-#[derive(Debug, Clone, Default)]
+#[derive(Debug, Clone)]
 pub struct DefaultMakeSpan {
+    level: Level,
     include_headers: bool,
 }
 
@@ -39,8 +42,19 @@ impl DefaultMakeSpan {
     /// Create a new `DefaultMakeSpan`.
     pub fn new() -> Self {
         Self {
+            level: DEFAULT_MESSAGE_LEVEL,
             include_headers: false,
         }
+    }
+
+    /// Set the [`Level`] used for the [tracing span].
+    ///
+    /// Defaults to [`Level::DEBUG`].
+    ///
+    /// [tracing span]: https://docs.rs/tracing/latest/tracing/#spans
+    pub fn level(mut self, level: Level) -> Self {
+        self.level = level;
+        self
     }
 
     /// Include request headers on the [`Span`].
@@ -54,23 +68,56 @@ impl DefaultMakeSpan {
     }
 }
 
+impl Default for DefaultMakeSpan {
+    fn default() -> Self {
+        Self::new()
+    }
+}
+
 impl<B> MakeSpan<B> for DefaultMakeSpan {
     fn make_span(&mut self, request: &Request<B>) -> Span {
-        if self.include_headers {
-            tracing::debug_span!(
-                "request",
-                method = %request.method(),
-                uri = %request.uri(),
-                version = ?request.version(),
-                headers = ?request.headers(),
-            )
-        } else {
-            tracing::debug_span!(
-                "request",
-                method = %request.method(),
-                uri = %request.uri(),
-                version = ?request.version(),
-            )
+        // This ugly macro is needed, unfortunately, because `tracing::span!`
+        // required the level argument to be static. Meaning we can't just pass
+        // `self.level`.
+        macro_rules! make_span {
+            ($level:expr) => {
+                if self.include_headers {
+                    tracing::span!(
+                        $level,
+                        "request",
+                        method = %request.method(),
+                        uri = %request.uri(),
+                        version = ?request.version(),
+                        headers = ?request.headers(),
+                    )
+                } else {
+                    tracing::span!(
+                        $level,
+                        "request",
+                        method = %request.method(),
+                        uri = %request.uri(),
+                        version = ?request.version(),
+                    )
+                }
+            }
+        }
+
+        match self.level {
+            Level::ERROR => {
+                make_span!(Level::ERROR)
+            }
+            Level::WARN => {
+                make_span!(Level::WARN)
+            }
+            Level::INFO => {
+                make_span!(Level::INFO)
+            }
+            Level::DEBUG => {
+                make_span!(Level::DEBUG)
+            }
+            Level::TRACE => {
+                make_span!(Level::TRACE)
+            }
         }
     }
 }